Replace spacecard component with spacedetailcard
